Added the version check script. added builds for the first round of packages.

+++ lfsbuild/task.py
@@ -2,12 +2,14 @@ import yaml
 import logging
 import os
 import glob
+import string
 
 import download
 import config
 
 log = logging.getLogger('TASK')
 
+tasks = {}
 
 class Task(object):
     """docstring for Task."""
@@ -21,6 +23,12 @@ class Task(object):
 
     def run(self):
         log.info('Running task: << {} >>'.format(self.name))
+
+        # check deps
+        if 'deps' in self.data:
+            for dep in self.data['deps']:
+                tasks[dep].run()
+
         url = self.data['src_uri']
         md5 = self.data['md5']
         filename = url[url.rfind('/')+1:]
@@ -35,23 +43,36 @@ class Task(object):
             tar_flags = 'xzf'
         elif url[-2:] == 'xz':
             tar_flags = 'xJf'
+        elif url[-3:] == 'bz2':
+            tar_flags = 'xjf'
         extract_cmd = 'tar {} {}/{} -C {}'.format(tar_flags, config.CACHE_PATH, filename, config.SOURCE_PATH)
-        log.info('Extracting {} to {}'.format(filename, config.SOURCE_PATH))
+        log.info('Extracting {} to {}'.format(filename, config.BUILD_PATH))
         return_code = os.system(extract_cmd)
         if return_code != 0:
             log.error('Error extracting {}'.format(filename))
             return False
 
-        # delete old build directory if it exists
-        if os.path.isdir(self.build_path):
-            os.system('rm -rf {}'.format(self.build_path))
+        if self.action == 'build':
+            # delete old build directory if it exists
+            if os.path.isdir(self.build_path):
+                os.system('rm -rf {}'.format(self.build_path))
 
-        # now create the build path and change to it
-        os.system('mkdir {}'.format(self.build_path))
-        os.chdir(self.build_path)
+            # check for premake_commands
+            if 'premake_commands' in self.data:
+                for pmc in self.data['premake_commands']:
+                    log.info('Running premake command: \n\r{}'.format(pmc))
+                    os.system(pmc)
 
-        os.system('../../{}/configure'.format(self.src_path))
-        os.system('make')
+            # now create the build path and change to it
+            os.system('mkdir {}'.format(self.build_path))
+            os.chdir(self.build_path)
+
+            configure_args = string.Template(self.config_opts).safe_substitute(config.CONFIGURE_VARS)
+            configure_cmd = '../../{}/configure {}'.format(self.src_path, configure_args)
+            log.info('Compiling {} with following options: {}'.format(self.name, configure_args))
+            os.system(configure_cmd)
+            os.system('make {}'.format(config.MAKE_OPTS))
+
     @property
     def name(self):
         return self.data['name']
@@ -64,9 +85,20 @@ class Task(object):
     def build_path(self):
         return '{}/{}'.format(config.BUILD_PATH, self.name)
 
+    @property
+    def config_opts(self):
+        return self.data['config_opts']
+
+    @property
+    def action(self):
+        try:
+            action = self.data['action']
+        except KeyError:
+            return 'none'
+        return action
+
 
 def load_tasks(path=config.TASK_PATH):
-    tasks = {}
     num_tasks = 0
     task_file_list = glob.glob('{}/*.yaml'.format(config.TASK_PATH))
     for f in task_file_list:
